requirements.txt
dj-database-url==0.5.0
Django==2.2.11
django-mdeditor==0.1.18
Markdown==3.2.1
python-decouple==3.3
sqlparse==0.3.1